0x01/mei/src/Person.java
import java.time.LocalDate;
import java.util.Date;



public class Person {

    private String name;
    private String surname;
    private Date birthDate;
    private boolean anotherCompanyOwner;
    private boolean pensioner;
    private boolean publicServer;
    private float salary;

    public static String fullName(String name, String surname) {
        return String.format("%s %s", name, surname);
    }

    public void setName(String name) {
        this.name = name;
    }

    public void setSurname(String surname) {
        this.surname = surname;
    }

    public Date getBirthDate() {
        return birthDate;
    }

    public void setBirthDate(Date birthDate) {
        this.birthDate = birthDate;
    }

    public boolean isAnotherCompanyOwner() {
        return anotherCompanyOwner;
    }

    public void setAnotherCompanyOwner(boolean anotherCompanyOwner) {
        this.anotherCompanyOwner = anotherCompanyOwner;
    }

    public boolean isPensioner() {
        return pensioner;
    }

    public void setPensioner(boolean pensioner) {
        this.pensioner = pensioner;
    }

    public boolean isPublicServer() {
        return publicServer;
    }

    public void setPublicServer(boolean publicServer) {
        this.publicServer = publicServer;
    }

    public float getSalary() {
        return salary;
    }

    public void setSalary(float salary) {
        this.salary = salary;
    }

    public static float calculateYearlySalary(float salary) {
        return salary * 12;
    }

    public static boolean isMEI(float salary, boolean anotherCompanyOwner, boolean pensioner, boolean publicServer, Date birthDate) {
        return (salary * 12) < 130000 && !anotherCompanyOwner && !pensioner && !publicServer && birthDate.getTime() < 2002;
    }
}